requirements.txt
fastapi
uvicorn
torch
torchvision
ftfy
regex
tqdm
pillow
clip-anytorch
psycopg2-binary
numpy
